Remove private imports from client, move util function only used by CLI to cli/_utils.py
