+++ vec_inf/client/__init__.py
@@ -5,6 +5,7 @@ using `vec_inf`. It is an alternative to the command-line interface, and allows
 users direct control over the lifecycle of inference servers via python scripts.
 """
 
+from vec_inf.client._config import ModelConfig
 from vec_inf.client._models import (
     LaunchOptions,
     LaunchOptionsDict,
@@ -28,4 +29,5 @@ __all__ = [
     "ModelType",
     "LaunchOptions",
     "LaunchOptionsDict",
+    "ModelConfig",
 ]
