+++ vec_inf/client/_utils.py
@@ -9,7 +9,6 @@ from typing import Any, Optional, Union, cast
 
 import requests
 import yaml
-from rich.table import Table
 
 from vec_inf.client._config import ModelConfig
 from vec_inf.client._models import ModelStatus
@@ -116,16 +115,6 @@ def model_health_check(
         return (ModelStatus.FAILED, response.status_code)
     except requests.exceptions.RequestException as e:
         return (ModelStatus.FAILED, str(e))
-
-
-def create_table(
-    key_title: str = "", value_title: str = "", show_header: bool = True
-) -> Table:
-    """Create a table for displaying model status."""
-    table = Table(show_header=show_header, header_style="bold magenta")
-    table.add_column(key_title, style="dim")
-    table.add_column(value_title)
-    return table
 
 
 def load_config() -> list[ModelConfig]:
